+++ VadereState/src/org/vadere/state/scenario/distribution/parameter/AttributesNormalDistribution.java
@@ -4,7 +4,8 @@ package org.vadere.state.scenario.distribution.parameter;
 import org.vadere.state.attributes.distributions.AttributesDistribution;
 
 /**
- * @author Lukas Gradl (lgradl@hm.edu)
+ * This is the parameter structure used with an normal distribution.
+ * @author Lukas Gradl (lgradl@hm.edu), Ludwig Jaeck
  */
 
 public class AttributesNormalDistribution extends AttributesDistribution {